Resturant Chain/requirements.txt
langchain
langchain_community
langchain_groq
python_dotenv
notebook